build_holoq_n1d7_pack: load the model in bfloat16 like the calibration server
The calibration artifact carries model_config_sha256 of the model the server
served, which Gr00tPolicy loads in bfloat16. The builder loaded with
AutoModel.from_pretrained defaults (float32), so config.dtype differed and the
builder refused every artifact with "Calibration config hash does not match
the loaded checkpoint". Load in bfloat16 so the statistics and the weights
they describe come from the same model; the guard itself is unchanged.

+++ tools/build_holoq_n1d7_pack.py
@@ -7,6 +7,8 @@ from __future__ import annotations
 
 import argparse
 
+import torch
+
 from gr00t.quantization.builder import build_holoq_pack
 from transformers import AutoModel
 
@@ -40,7 +42,10 @@ def main() -> None:
 
     import gr00t.model  # noqa: F401
 
-    model = AutoModel.from_pretrained(args.model_path).eval()
+    # Load exactly as Gr00tPolicy does (bfloat16): the calibration artifact records
+    # model_config_sha256 of the served model, and ``dtype`` is part of that config,
+    # so a default float32 load can never match the artifact.
+    model = AutoModel.from_pretrained(args.model_path, dtype=torch.bfloat16).eval()
     output = build_holoq_pack(
         model,
         calibration_path=args.calibration_path,
